Remove BaseRequest class and refactor code
The BaseRequest class was removed, impacting several other classes that extended it. As a result, these classes were refactored, moving some properties and annotations directly into them. Moreover, groupId property was removed from ThirdPartyApiConfiguration class and DroneImage file was moved to the inner package.
