+++ src/main/java/de/app/fivegla/controller/dto/request/ImageProcessingRequest.java
@@ -1,5 +1,6 @@
 package de.app.fivegla.controller.dto.request;
 
+import de.app.fivegla.controller.dto.request.inner.DroneImage;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -13,25 +14,19 @@ import java.util.List;
 @Getter
 @Setter
 @Schema(description = "Request for image processing.")
-public class ImageProcessingRequest extends BaseRequest {
+public class ImageProcessingRequest {
 
-    /**
-     * The transaction id.
-     */
     @NotBlank
     @Schema(description = "The transaction id.")
     private String transactionId;
 
-    /**
-     * The id of the drone.
-     */
     @NotBlank
     @Schema(description = "The id of the drone.")
     private String droneId;
 
-    /**
-     * The images to process.
-     */
     @Schema(description = "The images to process.")
     private List<DroneImage> images;
+
+    @Schema(description = "A custom group ID, which can be used to group devices / measurements. This is optional, if not set, the default group will be used.")
+    protected String groupId;
 }
