+++ src/main/java/de/app/fivegla/controller/dto/request/ImportAgriCropFromCsvRequest.java
@@ -7,18 +7,15 @@ import lombok.Setter;
 @Getter
 @Setter
 @Schema(description = "Request for importing agri-crop data from a CSV file.")
-public class ImportAgriCropFromCsvRequest extends BaseRequest {
+public class ImportAgriCropFromCsvRequest {
 
-    /**
-     * The ID of the crop.
-     */
     @Schema(description = "The ID of the crop.")
     private String cropId;
 
-    /**
-     * The CSV data.
-     */
     @Schema(description = "The CSV data.")
     private String csvData;
+
+    @Schema(description = "A custom group ID, which can be used to group devices / measurements. This is optional, if not set, the default group will be used.")
+    protected String groupId;
 
 }
