+++ src/main/java/de/app/fivegla/persistence/entity/ThirdPartyApiConfiguration.java
@@ -48,9 +48,4 @@ public class ThirdPartyApiConfiguration {
      */
     private String apiToken;
 
-    /**
-     * Represents the group for the third-party API.
-     */
-    private String groupId;
-
 }
