+++ src/main/resources/application.yml
@@ -7,7 +7,7 @@ app:
       # Spring Boot uses ISO-8601 durations, https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/time/Duration.html#parse(java.lang.CharSequence)
       initial-delay: PT10S # 10 seconds
       delay: PT30M # 30 minutes
-    daysInThePastForInitialImport: 3
+    daysInThePastForInitialImport: 1
   fiware:
     contextBrokerUrl: ${CONTEXT_BROKER_URL}
     subscriptions:
